[core] Continue work on improving apply specialization. (#3130)
We want to specialize and constant prop into apply callees. This
constant prop should be used to convert the specialized callee
into a DAG form. Also specialize any relaxed veq types while
we're at it.

Adjust pipelines and such on Ionq and Quantinuum to allow for
Grover example to JIT compile and execute.

Also tidy up the nvq++ driver a bit.

Add test.

Signed-off-by: Eric Schweitz <eschweitz@nvidia.com>